requirements.txt
django==1.10
MySQL-python==1.2.5
django-ckeditor==5.1.0
